fix error for merge

+++ dotnet/Repository/BrandRepository.cs
@@ -1,6 +1,7 @@
 using be_dotnet_ecommerce1.Data;
 using be_dotnet_ecommerce1.Model;
 using be_dotnet_ecommerce1.Repository.IRepository;
+using dotnet.Model;
 using Microsoft.EntityFrameworkCore;
 
 namespace be_dotnet_ecommerce1.Repository
@@ -17,10 +18,13 @@ namespace be_dotnet_ecommerce1.Repository
             return rs;
         }
 
-        public async Task<List<Brand>> getBrandByProductIds(List<int> ids)
+        public async Task<List<Brand>> getBrandByProductIds(List<long> ids)
         {
             var rs = await _connect.brands.Where(b => ids.Contains(b.id)).ToListAsync();
             return rs;
         }
+
+
+
     }
 }
